+++ explorer/ee/db_connections/admin.py
@@ -1,0 +1,9 @@
+from django.contrib import admin
+
+from explorer.models import DatabaseConnection
+
+
+@admin.register(DatabaseConnection)
+class DatabaseConnectionAdmin(admin.ModelAdmin):
+    pass
+