+++ explorer/src/js/main.js
@@ -15,7 +15,9 @@ const route_initializers = {
         new ExplorerEditor(document.getElementById('queryIdGlobal').value)),
     query_create:        () => import('./explorer').then(({ExplorerEditor}) => new ExplorerEditor('new')),
     explorer_playground: () => import('./explorer').then(({ExplorerEditor}) => new ExplorerEditor('new')),
-    explorer_schema:     () => import('./schema').then(({setupSchema}) => setupSchema())
+    explorer_schema:     () => import('./schema').then(({setupSchema}) => setupSchema()),
+    explorer_connection_create:            () => import('./uploads').then(({setupUploads}) => setupUploads()),
+    explorer_connection_update:            () => import('./uploads').then(({setupUploads}) => setupUploads())
 };
 
 document.addEventListener('DOMContentLoaded', function() {