+++ explorer/tests/test_utils.py
@@ -269,5 +269,5 @@ class TestConnections(TestCase):
 
         from explorer.app_settings import EXPLORER_DEFAULT_CONNECTION
         from explorer.connections import connections
-        self.assertTrue(EXPLORER_DEFAULT_CONNECTION in connections)
-        self.assertNotEqual(len(connections), len([c for c in djcs]))
+        self.assertTrue(EXPLORER_DEFAULT_CONNECTION in connections())
+        self.assertNotEqual(len(connections()), len([c for c in djcs]))